+++ threedigrid_builder/base/__init__.py
@@ -5,3 +5,4 @@ from .pumps import *  # NOQA
 from .surfaces import * # NOQA
 from .interfaces import *  # NOQA
 from .settings import *  # NOQA
+from .levees import *  # NOQA