+++ threedigrid_builder/constants.py
@@ -137,7 +137,6 @@ class InterflowType(IntEnum):
     LOCAL_DEEPEST_POINT_CONSTANT_POROSITY = 3
     GLOBAL_DEEPEST_POINT_CONSTANT_POROSITY = 4
 
-
 @unique
 class SurfaceClass(str, Enum):
     GESLOTEN_VERHARDING = "gesloten verharding"
@@ -145,3 +144,7 @@ class SurfaceClass(str, Enum):
     HALF_VERHARD = "half verhard"
     ONVERHARD = "onverhard"
     PAND = "pand"
+
+class Material(IntEnum):
+    SAND = 1
+    CLAY = 2