+++ threedigrid_builder/tests/test_obstacles.py
@@ -1,5 +1,6 @@
 from numpy.testing import assert_almost_equal
 from numpy.testing import assert_equal
+from threedigrid_builder.base import Levees
 from threedigrid_builder.base import Lines
 from threedigrid_builder.base import Nodes
 from threedigrid_builder.constants import LineType
@@ -57,7 +58,7 @@ def obstacles_no_intersect():
 
 
 def test_set_obstacles(grid, obstacles):
-    grid.set_obstacles(obstacles)
+    grid.set_obstacles(obstacles, Levees(id=[]))
 
     expected_kcu = np.array(
         [102, 102, 102, 102, 98, 99, 103, 99, 99, 103], dtype=np.int32
@@ -72,7 +73,7 @@ def test_set_obstacles(grid, obstacles):
 
 
 def test_set_obstacles_no_intersect(grid, obstacles_no_intersect):
-    grid.set_obstacles(obstacles_no_intersect)
+    grid.set_obstacles(obstacles_no_intersect, Levees(id=[]))
 
     assert np.all(grid.lines.kcu != LineType.LINE_2D_OBSTACLE)
     assert_equal(grid.lines.flod, np.nan)